requirements.txt
Flask
Flask-Cors
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-Testing
coverage
gunicorn
psycopg2
